Move stream aggregate state logic into Util
 - Refactoring updates

Resolves #2367
